Pulling

+++ CallejoIncChildcareAPI/Controllers/RoleController.cs
@@ -1,36 +1,51 @@
-﻿using Common.Services.Role;
+﻿using CallejoIncChildCareAPI.Authorize;
+using Common.Services.Login;
+using Common.Services.Role;
 using Common.View;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace CallejoIncChildcareAPI.Controllers
 {
-    [RequireHttps]
     [Route("api/[controller]")]
     [ApiController]
     public class RoleController : ControllerBase
     {
         private IRoleService _roleService;
+        private ILoginService _loginService;
 
         // Constructor to inject the IRoleService
-        public RoleController(IRoleService roleService)
+        public RoleController(IRoleService roleService,
+                              ILoginService loginService)
         {
             _roleService = roleService;
+            _loginService = loginService;
         }
 
 
         // GET: api/Role
+        [AuthorizeAttribute()]
         [HttpGet]
         public ActionResult<ListRoles> GetAllRoles()
         {
+            if (!_loginService.IsUserAuthenticated(AuthorizeAction.UserName, AuthorizeAction.AuthorizationToken))
+            {
+                return Unauthorized(new APIResponse { Success = false, Message = "User is not authenticated." });
+            }
             var result = _roleService.GetAllRoles();
             return Ok(result);
         }
 
         // GET: api/Role/{id}
+        [AuthorizeAttribute()]
         [HttpGet("{id}")]
         public ActionResult<ListRoles> GetRole(long id)
         {
+            if (!_loginService.IsUserAuthenticated(AuthorizeAction.UserName, AuthorizeAction.AuthorizationToken))
+            {
+                return Unauthorized(new APIResponse { Success = false, Message = "User is not authenticated." });
+            }
+
             var result = _roleService.GetRole(id);
             if (result.Success)
             {
@@ -40,9 +55,15 @@ namespace CallejoIncChildcareAPI.Controllers
         }
 
         // POST: api/Role
+        [AuthorizeAttribute()]
         [HttpPost]
         public ActionResult<APIResponse> InsertRole([FromBody] RoleView roleInfo)
         {
+            if (!_loginService.IsUserAuthenticated(AuthorizeAction.UserName, AuthorizeAction.AuthorizationToken))
+            {
+                return Unauthorized(new APIResponse { Success = false, Message = "User is not authenticated." });
+            }
+
             var result = _roleService.InsertRole(roleInfo);
             if (result.Success)
             {
@@ -52,9 +73,15 @@ namespace CallejoIncChildcareAPI.Controllers
         }
 
         // PUT: api/Role/{id}
+        [AuthorizeAttribute()]
         [HttpPut]
         public ActionResult<APIResponse> UpdateRole([FromBody] RoleView roleInfo)
         {
+            if (!_loginService.IsUserAuthenticated(AuthorizeAction.UserName, AuthorizeAction.AuthorizationToken))
+            {
+                return Unauthorized(new APIResponse { Success = false, Message = "User is not authenticated." });
+            }
+
             var result = _roleService.UpdateRole(roleInfo);
             if (result.Success)
             {
@@ -64,9 +91,15 @@ namespace CallejoIncChildcareAPI.Controllers
         }
 
         // DELETE: api/Role/{id}
+        [AuthorizeAttribute()]
         [HttpDelete("{id}")]
         public ActionResult<APIResponse> DeleteRole(long id)
         {
+            if (!_loginService.IsUserAuthenticated(AuthorizeAction.UserName, AuthorizeAction.AuthorizationToken))
+            {
+                return Unauthorized(new APIResponse { Success = false, Message = "User is not authenticated." });
+            }
+
             var result = _roleService.DeleteRole(id);
             if (result.Success)
             {
